requirements.txt
einops==0.8.2
evaluation==0.0.2
ftfy==6.3.1
hydra-core==1.3.2
numpy==1.23.5
omegaconf==2.3.0
packaging==26.0
Pillow==12.1.1
pytest==7.4.0
pytorch_lightning==2.6.1
regex==2024.11.6
regex==2023.10.3
setuptools==36.4.0
setuptools==75.8.0
setuptools==68.2.2
torch==2.0.0
torchvision==0.15.1+cu118
tqdm==4.67.1
tqdm==4.65.0
transformers==4.51.3
